Fixed narrowing conversion warning

+++ src/renderer/text_renderer.h
@@ -24,7 +24,7 @@ struct Character {
     unsigned int TextureID; // ID handle of the glyph texture
     glm::ivec2   Size;      // size of glyph
     glm::ivec2   Bearing;   // offset from baseline to left/top of glyph
-    unsigned int Advance;   // horizontal offset to advance to next glyph
+    long int Advance;   // horizontal offset to advance to next glyph
 };
